refactored code and moved errors to seperate folder
