+++ .gitignore
@@ -165,3 +165,4 @@ cython_debug/
 #.idea/
 
 toignore
+*_test.py
